requirements.txt
discord.py-self==1.9.1
discord.py==1.7.3
